remove const to file consts
